% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normHillCubes.R
\name{boolCubeFun}
\alias{boolCubeFun}
\title{asnormHillCube transform a boolean regulatory network to hill cube continuous interpolation.}
\usage{
boolCubeFun(boolInput, Input)
}
\arguments{
\item{net}{a regulatory network model that belongs to "booleanNetwork" or "squad" class.}

\item{n}{coefficient determines the level of the slope of the curve. A vector of
length k. Where k is the number of nodes of the network.}

\item{k}{coefficient determines the threshold which resembles boolean switch behavior.
A vector of length k. Where k is the number of nodes of the network.}

\item{gamma}{determines the dacay rate of the molecule. A vector of length k. Where k
is the number of nodes of the network.}
}
\description{
transform a boolean regulatory network to hill cube continuous interpolation.
It return a regulatory network model.
}
\examples{
library(BoolNet)
data(cellcycle)
cellcycle.ci <- asContinuous(cellcycle,to="normHillCubes")
initialState <- rep(0,length(cellcycle$genes))
sim <- squad(cellcycle.ci,initialState = initialState)
head(sim)

}
