% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFun.R, R/test.sq.R
\name{checkFun}
\alias{checkFun}
\alias{checkFun}
\title{test}
\usage{
test.sq(net, parameters)

test.sq(net, parameters)
}
\arguments{
\item{net}{a "SQUAD" or "BoolNet" object}

\item{net}{a "SQUAD" or "BoolNet" object}
}
\description{
function for testing proposes

function for testing proposes
}
\details{
function for testing proposes

function for testing proposes
}
