% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttractors.sq.R
\name{getAttractors.sq}
\alias{getAttractors.sq}
\title{Calculations of asynchronous steady state attractors of a Boolean Network}
\usage{
getAttractors.sq(net)
}
\arguments{
\item{net}{an object of class "BooleanNetwork" as that defined by the loadNetwork() function
of BoolNet R package}
}
\value{
a data frame with numerical binary values of the attractors
}
\description{
Finds all asynchronous stationary states of a Boolean Regulatory Network model. 
Takes as arguments an object of the "BoolNet" (or "SQUAD" under construction) and returns a 
data frame which contains the attractors of the model.
}
\details{
getAttractors.sq calculates all asynchronous attractors of a Boolean Regulatory Network model


The function is straightforward. It calls the ginsim library written by Naldi and 
colleagues, 2009 naldi2009logical.
}
\examples{
data(cellCycle) 
attractors <- getAttractors.sq(cellCycle)
attractors
}
