% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttractorsAsynchronous.R
\name{getAttractorsAsynchronous}
\alias{getAttractorsAsynchronous}
\title{getAttractors calculates all asynchronous attractors of a Boolean Regulatory Network model}
\usage{
getAttractorsAsynchronous()
}
\description{
getAttractors calculates all asynchronous attractors of a Boolean Regulatory Network model
}
