% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.sq.R
\name{heatmap.sq}
\alias{heatmap.sq}
\title{heatmap.sq draws a heatmap from the ODE simulations. In general, this function can plot any result
 derived from ode() function of the deSolve R Package}
\usage{
heatmap.sq(sim, indexes = "default")
}
\arguments{
\item{sim}{a trajectory of a continuous interpolation regulatory network model. A data frame as the result
of applying the ode() solver function.}

\item{indexes}{a vector of numerical or integer which corresponds to the node indexes are to be plotted.}
}
\description{
draw a heatmap of the simulation of a trajectory of the regulatory network using a
a continuous formalism.
}
\examples{
library(SQUAD)
data("cellcycle")
cellcycle.sq <- asContinuous(cellcycle)
sim <- squad(cellcycle.sq,initialState = generateState(cellcycle,specs = c("CycD"=1)))
heatmap.sq(sim,indexes = c(1,2,3,4))
}
