% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloJavaWorld.R
\name{helloJavaWorld}
\alias{helloJavaWorld}
\title{helloJavaWorld prints a hello Java World msj for testing purposes}
\usage{
helloJavaWorld()
}
\description{
helloJavaWorld prints a hello Java World msj for testing purposes
}
