% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNetwork.sq.R
\name{loadNetwork.sq}
\alias{loadNetwork.sq}
\alias{SQUAD}
\title{load a regulatory network model in squad format}
\usage{
SQUAD(x, w, gamma, h)
}
\arguments{
\item{fixed}{a vector of values in the range [0,1] to node values to be fixed.}
}
\description{
Load a regulatory network model in squad format. It returns a squad object that
can be used to perform continuous simulations using interpolations of Boolean Regulatory
network models
}
\examples{
net.sq <- loadNetwork.sq("cartoonNetworkSQUAD.txt")
sim <- squad(net.sq)
timeSerie.sq(sim)
}
