% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGraph.R
\name{randomMatrix}
\alias{randomMatrix}
\title{create a random conectivity matrix that can be used to test graphToModel() function}
\usage{
randomMatrix(numeric,integer)
}
\arguments{
\item{matrixDimension}{the number of rows (=columns) in the square matrix of connectivity}

\item{lamda}{a numeric positive value which represent the lamda parameter of the exponential distribution
for the number of outputs of each node.}
}
\description{
randomGraph() create a random conectivity matrix that can be used to test graphToModel() function
}
\examples{
randomMatrix(100,3)
str(randomMatrix)
}
