% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSerie.sq.R
\name{timeSerie.sq}
\alias{timeSerie.sq}
\title{timeSerie.sq plots regulatory network dynamic as a time serie}
\usage{
timeSerie.sq(dynamic, indexes = "default")
}
\arguments{
\item{dynamic}{a trajectory of a continuous interpolation regulatory network model. A data frame as the result
of applying the ode() or sqaud() solver function.}

\item{indexes}{a vector of numerical or integer which corresponds to the node indexes are to be plotted.}
}
\description{
timeSerie.sq plots regulatory network dynamic as a time serie.
}
\examples{
library(SQUAD)
data("cellcycle")
cellcycle.sq <- asContinuous(cellcycle)
sim <- squad(cellcycle.sq,initialState = generateState(cellcycle,specs = c("CycD"=1)))
timeSerie.sq(sim,indexes = c(1,2,3,4))
}
